+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased](https://github.com/axa-group/oauth2-mock-server/compare/v1.0.0...HEAD)
 
+### Added
+
+- Add Resource Owner Password Credentials grant
+
 ### Fixed
 
 - Add missing cache control headers on `/token` responses